requirements.txt
Django==2.1.7
djangorestframework==3.9.2
pkg-resources==0.0.0
pytz==2018.9
qrcode==6.1
six==1.12.0
